# user-service/openapi.json
{
  "openapi": "3.0.0",
  "info": { "title": "User Service API", "version": "1.0.0", "description": "Authentication, profile and cart APIs exposed through API Gateway." },
  "servers": [{ "url": "http://localhost:8000/api/users", "description": "API Gateway route" }],
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } }
  },
  "paths": {
...
    "/register": { "post": { "summary": "Register a new user", "tags": ["Auth"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["email", "password", "full_name"], "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string", "minLength": 6 }, "full_name": { "type": "string" } } } } } }, "responses": { "201": { "description": "Registered" }, "400": { "description": "Invalid input" } } } },
    "/login": { "post": { "summary": "Login and receive JWT", "tags": ["Auth"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["email", "password"], "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string" } } } } } }, "responses": { "200": { "description": "Logged in" }, "401": { "description": "Invalid credentials" } } } },
    "/profile": { "get": { "summary": "Get current profile", "tags": ["Profile"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Profile data" }, "401": { "description": "Unauthorized" } } }, "put": { "summary": "Update current profile", "tags": ["Profile"], "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["full_name"], "properties": { "full_name": { "type": "string" } } } } } }, "responses": { "200": { "description": "Updated" }, "400": { "description": "Invalid input" } } } },
    "/change-password": { "put": { "summary": "Change current password", "tags": ["Profile"], "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["currentPassword", "newPassword"], "properties": { "currentPassword": { "type": "string" }, "newPassword": { "type": "string", "minLength": 6 } } } } } }, "responses": { "200": { "description": "Password changed" }, "400": { "description": "Invalid input" } } } },
    "/cart": { "get": { "summary": "Get current user's cart", "tags": ["Cart"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Cart data" }, "401": { "description": "Unauthorized" } } } },
    "/cart/add": { "post": { "summary": "Add item to cart", "tags": ["Cart"], "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["productId", "quantity", "color", "size"], "properties": { "productId": { "type": "string" }, "quantity": { "type": "integer", "minimum": 1 }, "color": { "type": "string" }, "size": { "type": "integer" } } } } } }, "responses": { "200": { "description": "Added" }, "400": { "description": "Invalid input" } } } },
    "/cart/update": { "put": { "summary": "Update cart item quantity", "tags": ["Cart"], "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["cartItemId", "quantity"], "properties": { "cartItemId": { "type": "integer" }, "quantity": { "type": "integer", "minimum": 1 } } } } } }, "responses": { "200": { "description": "Updated" }, "400": { "description": "Invalid input" } } } },
    "/cart/remove/{cartItemId}": { "delete": { "summary": "Remove cart item", "tags": ["Cart"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "cartItemId", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Removed" }, "404": { "description": "Not found" } } } },
    "/cart/clear": { "delete": { "summary": "Clear current user's cart", "tags": ["Cart"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Cleared" } } } }
  }
}